Message writter tests
