+++ photonpump/connection.py
@@ -406,43 +406,6 @@ class MessageDispatcher:
                 await sub.events.enqueue(StopIteration())
 
 
-class MessageWriter:
-
-    def __init__(self, queue, connector):
-        connector.conneted += self.on_connected
-        connector.on_connection_lost += self.on_connection_lost
-        self._queue = queue
-
-    def on_connection_lost(self):
-        self._write_loop.cancel()
-
-    def on_connected(self, _, streamwriter):
-        self.stream_wrtier = streamwriter
-        self.write_loop = asyncio.ensure_future(
-            self._write_outbound_messages()
-        )
-
-    async def _write_outbound_messages(self):
-        if self.next:
-            self.stream_wrtier.write(self.next.header_bytes)
-            self.stream_wrtier.write(self.next.payload)
-
-        while self._is_connected:
-            self._logger.debug('Sending message %s', self.next)
-            self.next = await self._queue.get()
-            try:
-                self.stream_wrtier.write(self.next.header_bytes)
-                self.stream_wrtier.write(self.next.payload)
-            except Exception as e:
-                self._logger.error(
-                    'Failed to send message %s', e, exc_info=True
-                )
-            try:
-                await self.stream_wrtier.drain()
-            except Exception as e:
-                self._logger.error(e)
-
-
 class PersistentSubscription(convo.PersistentSubscription):
 
     def __init__(self, subscription, iterator, conn, out_queue=None):
