+++ test/connection/fake_server.py
@@ -1,5 +1,7 @@
 import asyncio
 
+from photonpump.connection2 import Event
+
 
 class FakeProtocol(asyncio.Protocol):
